test: refactor, fix and adjust pr promotion tests
